+++ Cython/Compiler/Future.py
@@ -10,5 +10,6 @@ print_function = _get_feature("print_function")
 absolute_import = _get_feature("absolute_import")
 nested_scopes = _get_feature("nested_scopes")  # dummy
 generators = _get_feature("generators")  # dummy
+generator_stop = _get_feature("generator_stop")
 
 del _get_feature